Rebased `resolveAvroTypeNullability` util to `AvroConversionUtils`
